Implemented arithmetic and logic instructions

+++ src/Device.java
@@ -2,5 +2,5 @@ public interface Device {
     public boolean hasAddress(int address);
     public boolean IO_Mbar();
     public void writeByte(int address, int data);
-    public short readByte(int address);
+    public int readByte(int address);
 }
